+++ test/pharma/pharma_create_db_schema.py
@@ -1,11 +1,10 @@
-# Copyright IBM All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
+# Creates a schema in the DB
+# WARNING: will delete all existing tables
 import os
 
-# from dashboard.my_secrets.db2wh import DB2Cloud_DO_Dashboards_credentials
-from fruit_dash_app import FruitDashApp
+from pharma.pharma_dash_app import PharmaDashApp
 from dse_do_dashboard.dash_app import HostEnvironment
+from supply_chain.pharma.pharmascenariodbtables import PharmaScenarioDbManager
 
 if 'PROJECT_NAME' in os.environ:  # This works in CP4D v4.0.2
     host_env = HostEnvironment.CPD402  #'CP4D'
@@ -17,13 +16,5 @@ else:
     from my_secrets.db2wh import DB2Cloud_DO_Dashboards_credentials
     DB2_credentials = DB2Cloud_DO_Dashboards_credentials
 
-DA = FruitDashApp(db_credentials=DB2_credentials, schema='FRUIT_V2', dash_debug=True, host_env=host_env,
-                  #                   port=8051,
-                  db_echo=True,
-                  )
-
-
-
-####################################
-if __name__ == '__main__':
-    DA.run_server()
+scdb = PharmaScenarioDbManager(credentials = DB2_credentials, schema='PHARMA_V1', echo=True)
+scdb.create_schema()