+++ test/pharma/supply_chain/pharma/pharmascenariodbtables.py
@@ -17,6 +17,8 @@ from supply_chain.scnfo.scnfoscenariodbtables import ProductTable, ScnfoScenario
 from dse_do_utils.scenariodbmanager import ScenarioDbTable, AutoScenarioDbTable  #, ScenarioDbManager
 # from supply_chain.water.dewaterdatamanager import DEWaterDataManager  #ScenarioDbTable, ScenarioDbManager
 
+
+
 import pandas as pd
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++