+++ test/pharma/supply_chain/scnfo/scnfoscenariodbtables.py
@@ -11,7 +11,7 @@ from dse_do_utils.scenariodbmanager import ScenarioDbTable, ScenarioDbManager
 # from supply_chain.supply_chain import DEWaterDataManager  #ScenarioDbTable, ScenarioDbManager
 
 import pandas as pd
-
+from dse_do_dashboard.utils.scenariodbmanager_update import ScenarioDbManagerUpdate
 
 class ScenarioTable(ScenarioDbTable):
     def __init__(self, db_table_name: str = 'scenario'):
@@ -526,7 +526,7 @@ class BusinessKpiTable(ScenarioDbTable):
         columns_metadata.extend(extended_columns_metadata)
         super().__init__(db_table_name, columns_metadata)
         
-class ScnfoScenarioDbManager(ScenarioDbManager):
+class ScnfoScenarioDbManager(ScenarioDbManagerUpdate):
     def __init__(self, input_db_tables: Dict[str, ScenarioDbTable]=None, output_db_tables: Dict[str, ScenarioDbTable]=None, 
                  credentials=None, schema: str = None, echo=False, multi_scenario: bool = True):
         if input_db_tables is None: